guidance_agent/requirements.txt
absl-py==1.4.0
accelerate==0.19.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
altair==4.2.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.38.1
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.0
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
chromadb==0.3.23
click==8.1.3
clickhouse-connect==0.5.24
clip @ git+https://github.com/openai/CLIP.git@3702849800aa56e2223035bccd1c6ef91c704ca8
cmake==3.26.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dadaptation==1.5
dataclasses-json==0.5.7
datasets==2.12.0
dearpygui==1.8.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.10.2
dill==0.3.6
discord==2.2.3
discord.py==2.2.3
diskcache==5.6.1
duckdb==0.8.0
easygui==0.98.3
einops==0.6.0
entrypoints==0.4
executing==1.2.0
fairscale==0.4.13
fastapi==0.95.0
fastjsonschema==2.17.1
ffmpy==0.3.0
filelock==3.12.0
fire==0.5.0
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==23.3.3
flexgen==0.1.7
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
gast==0.4.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.25.0
gradio_client==0.1.4
greenlet==2.0.2
grpcio==1.51.3
guidance==0.0.55
h11==0.14.0
h5py==3.8.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.14.1
humanize==4.6.0
idna==2.10
imageio==2.26.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
InstructorEmbedding==1.0.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langchain==0.0.176
lazy_loader==0.2
libclang==16.0.0
library @ file:///home/karajan/labzone/2D/kohya_ss
lightning-utilities==0.8.0
linkify-it-py==2.0.0
lion-pytorch==0.0.6
lit==16.0.3
llama-cpp-python==0.1.45
lycoris-lora==0.1.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions @ file:///home/conda/feedstock_root/build_artifacts/mypy_extensions_1675543315189/work
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
notion-client==2.0.0
notion-sdk-py==0.0.1
numexpr==2.8.4
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai==0.27.7
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
orjson==3.8.12
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parsimonious==0.10.0
parso==0.8.3
peft @ git+https://github.com/huggingface/peft@b1059b73aab9043b118ff19b0cf96263ea86248a
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
-e git+https://github.com/openai/point-e.git@fc8a607c08a3ea804cc82bf1ef8628f88a3a5d2f#egg=point_e
posthog==3.0.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
PuLP==2.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.8
pydub==0.25.1
Pygments==2.14.0
pygtrie==2.5.0
PyMCubes==0.1.4
pymeshlab==2022.2.post2
pyparsing==3.0.9
pyre-extensions @ file:///home/conda/feedstock_root/build_artifacts/pyre-extensions_1677079251428/work
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
quant-cuda==0.0.0
qudida==0.0.4
regex==2023.5.5
requests==2.25.1
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.1
rsa==4.9
rwkv==0.7.3
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
stack-data==0.6.2
starlette==0.26.1
sympy==1.12
tenacity==8.2.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.3.21
tiktoken==0.4.0
timm==0.6.12
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==1.13.0+cu117
torch-ema==0.3
torchmetrics==0.11.4
torchvision==0.14.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
trimesh==3.20.1
triton==2.0.0.post1
typing-inspect @ file:///home/conda/feedstock_root/build_artifacts/typing_inspect_1674105119082/work
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
uri-template==1.2.0
urllib3==1.26.4
uvicorn==0.21.1
uvloop==0.17.0
voluptuous==0.13.1
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==10.4
Werkzeug==2.3.4
widgetsnbextension==4.0.7
wrapt==1.15.0
xatlas==0.0.7
xformers @ file:///home/conda/feedstock_root/build_artifacts/xformers_1680932431165/work
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
zstandard==0.21.0
